Add initial register and bit macros headers
Adds header files to define register addresses for the RCC and GPIOD
peripherals, and modifies the blink project to use these definitions.
More header files for peripheral registers will be added later as they
are needed. This allows register definitions to be resused by multiple
projects. Also, makes it easier to adapt these projects to different
hardware, as register definitions for different hardware targets can be
added.

The hardware-files folder is renamed to stm32f407vg, since this is the
hardware used by this repo.

Modifies the blink project to include these new files, and removes the
register and macro definitions from that project.